Part 1 - Inheritance

+++ Inheritance/SalariedEmployee.java
@@ -1,22 +1,22 @@
 package Inheritance;
 public class SalariedEmployee extends Employee {
-    double weekly_salary;
+    int weekly_salary;
 
-    public SalariedEmployee(String f, String l, int sn, double ws) {
+    public SalariedEmployee(String f, String l, int sn, int ws) {
         super(f, l, sn);
         this.weekly_salary = ws;
     }
 
-    public void setWeeklySalary(double ws) {
+    public void setWeeklySalary(int ws) {
         this.weekly_salary = ws;
     }
 
-    public double getWeeklySalary() {
+    public int getWeeklySalary() {
         return this.weekly_salary;
     }
 
     @Override
     public String toString() {
-        return super.toString() + String.format(" | weekly salary: $%.2f", this.weekly_salary);
+        return super.toString() + String.format(" | weekly salary: $%d", this.weekly_salary);
     }
 }
